Safari-609.1.11.1/LayoutTests/css3/flexbox/percentage-max-width-cross-axis.html

<!DOCTYPE html>

<link rel="stylesheet" href="resources/flexbox.css">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<body onload="checkLayout('.flexbox')">

<p>Both bars should be the same width</p>
<div class="flexbox column" style="width: 400px;">
  <div data-expected-width="200" style="margin-left: 50%; width: 50%; background: green;">
    Content
  </div>
</div>

<div class="flexbox column" style="width: 400px;">
  <div data-expected-width="200" style="margin-left: 50%; max-width: 50%; background: blue;">
    Content
  </div>
</div>
